Better TeeZip outtype test

+++ test/test_library.jl
@@ -94,8 +94,8 @@ end
             collect(zip(1:2:5, 2:2:6))
     end
 
-    ed = eduction(TeeZip(Filter(isodd) |> Map(x -> x + 1)), 1:5)
-    @test eltype(ed) === Tuple{Int,Int}
+    ed = eduction(TeeZip(Filter(isodd) |> Map(x -> x + 1.0)), 1:5)
+    @test eltype(ed) === Tuple{Int,Float64}
 
     xf = Map(inc) |> TeeZip(Filter(isodd)) |> Map(first)
     @testset for xs in iterator_variants(1:6)
